FIX: Balls keep crossing the outer boundaries of the game space
I don't want to fixate that if any gameObject leaves the drawable area that they are removed automatically. So, I decided to simply increase the outer boundaries so that they do not cross over to other side. If this somehow is enough to solve this bug, I will probably never touch it again